requirements.txt
aiogram>=3.12.0
fastapi
uvicorn
openai
shodan
duckduckgo_search
google-search-results
apscheduler
psutil
python-dotenv
fpdf2
SQLAlchemy
aiosqlite
asyncpg
redis
python-whois
dnspython
qrcode
pillow
